refactor: extract service
Extract packument-fetch service
